Removed the iris specific code from student.library
