+++ iris/src/main/java/tds/iris/web/data/ContentRequestItem.java
@@ -1,0 +1,32 @@
+/*******************************************************************************
+ * Educational Online Test Delivery System 
+ * Copyright (c) 2014 American Institutes for Research
+ *       
+ * Distributed under the AIR Open Source License, Version 1.0 
+ * See accompanying file AIR-License-1_0.txt or at
+ * http://www.smarterapp.org/documents/American_Institutes_for_Research_Open_Source_Software_License.pdf
+ ******************************************************************************/
+package tds.iris.web.data;
+
+import tds.itemrenderer.data.ITSTypes.ITSEntityType;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class ContentRequestItem extends ContentRequestEntity
+{
+  private String _response;
+
+  public ContentRequestItem () {
+    super ();
+    setItsType (ITSEntityType.Item);
+  }
+
+  @JsonProperty ("response")
+  public String getResponse () {
+    return _response;
+  }
+
+  public void setResponse (String value) {
+    _response = value;
+  }
+}
