+++ pom.xml
@@ -9,6 +9,7 @@
     <version>2.1.2-SNAPSHOT</version>
 
     <modules>
+        <module>student-shared</module>
         <module>student</module>
         <module>student.library</module>
         <module>iris</module>
