+++ student-shared/pom.xml
@@ -1,0 +1,43 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>org.opentestsystem.shared</groupId>
+        <artifactId>shared-master</artifactId>
+        <version>1.0.1-BUILD-SNAPSHOT</version>
+    </parent>
+    <groupId>org.opentestsystem.delivery</groupId>
+    <artifactId>student-shared</artifactId>
+    <version>0.0.1-SNAPSHOT</version>
+    <packaging>jar</packaging>
+
+    <name>student-shared</name>
+    <description>Contains shared libraries across projects</description>
+    <url>http://maven.apache.org</url>
+
+    <properties>
+        <resource-bundler.version>0.0.1-SNAPSHOT</resource-bundler.version>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+
+        <item-selection.version>0.0.1-SNAPSHOT</item-selection.version>
+        <item-renderer.version>0.0.1-SNAPSHOT</item-renderer.version>
+    </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.opentestsystem.delivery</groupId>
+            <artifactId>item-renderer</artifactId>
+            <version>${item-renderer.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>javax.servlet</groupId>
+            <artifactId>servlet-api</artifactId>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>AIR</groupId>
+            <artifactId>ResourceBundler</artifactId>
+            <version>${resource-bundler.version}</version>
+        </dependency>
+    </dependencies>
+</project>
