--- student.library/src/main/java/tds/iris/web/data/ContentRequestPassage.java
@@ -1,32 +1,0 @@
-/*******************************************************************************
- * Educational Online Test Delivery System 
- * Copyright (c) 2014 American Institutes for Research
- *       
- * Distributed under the AIR Open Source License, Version 1.0 
- * See accompanying file AIR-License-1_0.txt or at
- * http://www.smarterapp.org/documents/American_Institutes_for_Research_Open_Source_Software_License.pdf
- ******************************************************************************/
-package tds.iris.web.data;
-
-import tds.itemrenderer.data.ITSTypes.ITSEntityType;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class ContentRequestPassage extends ContentRequestEntity
-{
-  private boolean _autoLoad = true;
-
-  public ContentRequestPassage () {
-    super ();
-    setItsType (ITSEntityType.Passage);
-  }
-
-  @JsonProperty ("autoLoad")
-  public boolean getAutoLoad () {
-    return _autoLoad;
-  }
-
-  public void setAutoLoad (boolean value) {
-    _autoLoad = value;
-  }
-}
